154 hadolint (#155)
* hadolint in pre-commit

* Add pytest to test requirements

Include pytest in `requirements.txt` to enable testing functionality. This ensures tests can be executed consistently across environments.

* Refactor Dockerfile to include multi-stage builds.

Introduced separate stages for production and testing to optimize the Docker image. The `prod` stage focuses on running the application, while the `Test` stage includes test dependencies and configurations for easier test execution. This change improves security, efficiency, and maintainability of the Docker build process.

* Update release workflow to use Podman and Buildah actions

Replaced Docker-based build and push steps with Podman and Buildah actions for better container management and registry compatibility. Images are now built with Buildah and pushed to GitHub Container Registry.

* Update Docker build to use 'prod' target

Changed the Docker build command to include the '--target prod' option. This ensures the build process uses the production stage, enhancing consistency with production deployments.

* Add workflow to test Docker image build and execution

This new GitHub Actions workflow builds a Docker image targeting the "test" stage of the Dockerfile and runs it to ensure functionality. It triggers on every push and uses `podman` for execution, enhancing CI/CD pipelines for the repository.

* Fix COPY command syntax in Dockerfile

Corrected the COPY command to include a trailing slash for the destination directory. This ensures proper handling of the directory structure during the build process and avoids potential build errors.

* Rename Docker build stage "Test" to lowercase "test"

This change standardizes the naming convention for Docker build stages by using lowercase for consistency. Aligning with common practices improves readability and maintainability.